Place headers into `gapbs` namespace
Placing all the headers into the `gapbs` namespace so that they may be
used in other projects without polluting the global namespace. I also
had to move the `optarg` declaration to the global scope in
`command_line.h` to avoid linking issues. Basically, `optarg` is
implicitly defined, so the namespace can't be changed.