Removing ASA from gap code
Additionally adding hook to graph class that allows returns the
key,value pair for a given vertex. Note that since we use
`std::lower_bound` it is not "safe", but will return a garbage value
even if the vertex (u,v) is not in the graph (checking is left to the
user).