Updating Shiloach-Vishkin Connected Components Signature
Changing the loop range to be over a user specified range. Note
that this range should identify a seqeuence of vertices to be
computed on. We've also correspondingly changed `comp` to be
passed in as a parameter which is written two. In the case,
of a non-contiguous range, the components will only be updated
according to which vertices are associated with the range.